Display Loaded Skylanders in Overlay and Qt
